+++ src/components/HorizontalBarChart/hooks/index.ts
@@ -1,3 +1,0 @@
-export {useBarSizes} from './useBarSizes';
-export {useDataForChart} from './useDataForChart';
-export {useXScale} from './useXScale';