+++ src/components/shared/HorizontalStackedBars/index.ts
@@ -1,0 +1,1 @@
+export {HorizontalStackedBars} from './HorizontalStackedBars';