+++ src/components/shared/index.ts
@@ -1,0 +1,6 @@
+export {GradientDefs, getGradientDefId} from './GradientDefs';
+export {GroupLabel} from './GroupLabel';
+export {Bar} from './Bar';
+export {HorizontalBars} from './HorizontalBars';
+export {HorizontalStackedBars} from './HorizontalStackedBars';
+export {HorizontalGroup} from './HorizontalGroup';