+++ src/utilities/index.ts
@@ -25,3 +25,6 @@ export {
   changeGradientOpacity,
 } from './change-color-opacity';
 export {getRoundedRectPath} from './get-rounded-rect-path';
+export {getBarId} from './get-bar-id';
+export {formatDataForHorizontalBarChart} from './format-data-for-horizontal-bar-chart';
+export {getHighestSumForStacked} from './get-highest-sum-for-stacked';